Moving functions to the groups
